refactor: migrate collaborator api
