+++ packages/insomnia-api/src/index.ts
@@ -3,5 +3,6 @@ export * from './vault';
 export * from './enterprise';
 export * from './trial';
 export * from './project';
+export * from './collaborators';
 
 export { configureFetch, type FetchConfig, ResponseFailError, isApiError } from './fetch';
