+++ packages/insomnia/src/routes/organization.$organizationId.collaborators-search.tsx
@@ -1,19 +1,10 @@
+import { searchCollaborators } from 'insomnia-api';
 import { href } from 'react-router';
 
 import { userSession } from '~/models';
-import { insomniaFetch } from '~/ui/insomnia-fetch';
 import { createFetcherLoadHook } from '~/utils/router';
 
 import type { Route } from './+types/organization.$organizationId.collaborators-search';
-
-type CollaboratorType = 'invite' | 'member' | 'group';
-
-interface CollaboratorSearchResultItem {
-  id: string;
-  picture: string;
-  type: CollaboratorType;
-  name: string;
-}
 
 export async function clientLoader({ params, request }: Route.ClientLoaderArgs) {
   const { id: sessionId } = await userSession.get();
@@ -24,10 +15,10 @@ export async function clientLoader({ params, request }: Route.ClientLoaderArgs) 
     const requestUrl = new URL(request.url);
     const searchParams = Object.fromEntries(requestUrl.searchParams.entries());
 
-    const collaboratorsSearchList = await insomniaFetch<CollaboratorSearchResultItem[]>({
-      method: 'GET',
-      path: `/v1/desktop/organizations/${organizationId}/collaborators/search/${searchParams.query}`,
+    const collaboratorsSearchList = await searchCollaborators({
       sessionId,
+      organizationId,
+      keyword: searchParams.query || '',
     });
 
     return collaboratorsSearchList;
